requirements.txt
flask==2.3.3
flask-pymongo==2.3.0
python-dotenv==1.0.0
flask-bcrypt==1.0.1
flask-cors==4.0.0
PyJWT==2.8.0
flask-restful==0.3.10
marshmallow==3.20.1
gunicorn==21.2.0
python-json-logger==2.0.7